chore: add sbom generation and upload workflow
Adds stand-alone workflow to automatically generate and publish an SBOM
following a GitHub release (release event).

The workflow can also be triggered manually (workflow_dispatch event)
for testing, or to generate SBOMs for previous releases.

**Caveat:**
When creating an SBOM for a previous release, the workflow checks out
the repository at the corresponding tag. This means that the used SBOM
schema version, bumped to 1.6 in the current patch, will be
reverted back to 1.4.

**Recommendation:**
Move full SBOM configuration from pom.xml to the command line invocation
in the workflow file added in this patch, to avoid surprising behaviour.
NOTE: The existing execution goal in pom.xml does not seem to work (see
"Skipping CycloneDX goal, because module skips deploy" in build logs of
recent releases).

Please reach out to the security team for assistance!

Signed-off-by: Lukas Puehringer <lukas.puehringer@eclipse-foundation.org>